+++ custom_model_runner/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##### Fixed
 - Fixed issue with multiclass scoring failing with numerical class labels
+- Fixed multclass class labels file not working in --docker mode
 
 #### [1.4.1] - 2020-10-29
 ##### Added